+++ src/redux/epics.js
@@ -7,6 +7,7 @@ import * as supervisorToolbarEpics from './modules/supervisorToolbar/epics';
 import * as emailTemplatesEpics from './modules/entities/emailTemplates/epics';
 import * as usersEpics from './modules/entities/users/epics';
 import * as dataAccessReportsEpics from './modules/entities/dataAccessReports/epics';
+import * as reasonListsEpics from './modules/entities/reasonLists/epics';
 
 export const rootEpic = combineEpics(
   ...Object.values(brandingEpics),
@@ -16,5 +17,6 @@ export const rootEpic = combineEpics(
   ...Object.values(supervisorToolbarEpics),
   ...Object.values(emailTemplatesEpics),
   ...Object.values(usersEpics),
-  ...Object.values(dataAccessReportsEpics)
+  ...Object.values(dataAccessReportsEpics),
+  ...Object.values(reasonListsEpics)
 );