+++ src/redux/modules/entities/config.js
@@ -1,4 +1,4 @@
 export const hasCustomCreateEntity = entityName => !['users', 'dataAccessReports'].includes(entityName);
 
 export const hasCustomUpdateEntity = entityName =>
-  !['emailTemplates', 'users', 'dataAccessReports'].includes(entityName);
+  !['emailTemplates', 'users', 'dataAccessReports','reasonLists'].includes(entityName);