+++ src/redux/modules/entities/reasonLists/__tests__/selectors.test.js
@@ -4,9 +4,11 @@
 
 import { fromJS } from 'immutable';
 import { reasonListsInitialValues, checkDisableShared } from '../selectors';
+import { getSelectedEntity, getSelectedEntityId } from '../../selectors';
 import { selectFormInitialValues } from '../../../form/selectors';
 
 jest.mock('../../../form/selectors');
+jest.mock('../../selectors');
 
 describe('reasonListsInitialValues', () => {
   it('Return the initial Reason List Values, all values provided', () => {
@@ -14,9 +16,35 @@ describe('reasonListsInitialValues', () => {
       id: 'mockId',
       shared: false,
       isDefault: false,
+      reasons: [
+        {
+          reasonId: 'mockReasonId',
+          sortOrder: '0',
+          hierarchy: []
+        },
+        {
+          reasonId: 'mockReasonId2',
+          sortOrder: '1',
+          hierarchy: []
+        }
+      ]
+    });
+    selectFormInitialValues.mockImplementation(() => initialState);
+    getSelectedEntity.mockImplementation(() => initialState);
+    getSelectedEntityId.mockImplementation(() => 'mockId');
+    expect(reasonListsInitialValues()).toMatchSnapshot();
+  });
+
+  it('Empty Reason List, all values provided', () => {
+    const initialState = fromJS({
+      id: 'mockId',
+      shared: false,
+      isDefault: false,
       reasons: []
     });
     selectFormInitialValues.mockImplementation(() => initialState);
+    getSelectedEntity.mockImplementation(() => initialState);
+    getSelectedEntityId.mockImplementation(() => 'mockId');
     expect(reasonListsInitialValues()).toMatchSnapshot();
   });
 
@@ -50,5 +78,29 @@ describe('checkDisableShared', () => {
     });
     selectFormInitialValues.mockImplementation(() => initialState);
     expect(checkDisableShared()).toMatchSnapshot();
+  });
+});
+
+describe('filterReasons', () => {
+  it('Return Filtered Reasons, all values provided', () => {
+    const initialState = fromJS({
+      id: 'mockId',
+      shared: false,
+      isDefault: false,
+      reasons: [
+        {
+          reasonId: 'mockReasonId',
+          sortOrder: '0',
+          hierarchy: []
+        },
+        {
+          reasonId: 'mockReasonId2',
+          sortOrder: '1',
+          hierarchy: []
+        }
+      ]
+    });
+    getSelectedEntity.mockImplementation(() => initialState);
+    expect(reasonListsInitialValues()).toMatchSnapshot();
   });
 });